% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreInvHap.R
\name{adaptRefs}
\alias{adaptRefs}
\title{Adapt references to imputed data}
\usage{
adaptRefs(Refs, alleletable, haploid = FALSE)
}
\arguments{
\item{Refs}{List with the allele frequencies}

\item{alleletable}{Data.frame with the alleles per SNP (from getAlleleTable)}

\item{haploid}{Logical. If TRUE, modify references for haploid samples}
}
\value{
List with the same values than Refs but adapted to imputation data
}
\description{
Internal
}
