% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierPipeline.R
\name{classifierPipeline}
\alias{classifierPipeline}
\title{Classify inversions using identity}
\usage{
classifierPipeline(SNPlist, SNPsR2, hetRefs, Refs, ldamodel, R2 = 0,
  uncertainy.thres = 0.99, mc.cores = 1, verbose = FALSE)
}
\arguments{
\item{SNPlist}{List with SNPs data. It should contain genotypes (a SNPmatrix) and map (a data.frame
with the annotation)}

\item{SNPsR2}{Vector with the R2 of the SNPs of the region}

\item{hetRefs}{Vector with the heterozygote form of the SNP in the inversion}

\item{Refs}{List with the allele frequencies in the references}

\item{ldamodel}{LDAmodel used to classify individuals}

\item{R2}{Vector with the R2 between the SNPs and the inversion status}

\item{uncertainy.thres}{Numeric with the uncertainty threshold. If the maximum certainty of belonging
to a group is smaller than this parameter, the individual is classified as NA.}

\item{mc.cores}{Numeric with the number of cores used in the computation}

\item{verbose}{Should message be shown?}
}
\value{
A list with the following elements:
\itemize{
\item{class: Factor with the classification of the individuals}
\item{scores: Matrix with the simmilarity scores of the individuals}
}
}
\description{
Classify inversions using identity
}

