% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifSNPs.R
\name{classifSNPs}
\alias{classifSNPs}
\title{Get similarity scores and probability}
\usage{
classifSNPs(genos, R2, refs, alfreq, genofreq, mc.cores)
}
\arguments{
\item{genos}{Matrix with the samples genotypes. It is the result of \code{getGenotypesTable}}

\item{R2}{Vector with the R2 between the SNPs and the inversion status}

\item{refs}{List of matrices. Each matrix has, for an SNP, the frequencies of each genotype in the
different haplotypes.}

\item{alfreq}{List with the allele frequencies in the target population.}

\item{genofreq}{Vector with the frequency of each haplotype}

\item{mc.cores}{Numeric with the number of cores used in the computation}
}
\value{
List with the results:
\itemize{
\item{scores: Matrix with the simmilarity scores of the individuals}
\item{probs: Matrix with the probability of each individual to belong to the
different haplotypes}
\item{numSNPs: Vector with the number of SNPs used in each computation}
}
}
\description{
This function computes the similarity scores between the sample SNPs and the haplotype's
reference. It also returns the probability of each sample to belong to the
different haplotypes.
}
\details{
This function computes, for each individual, similarity scores for
all the present haplotypes. For each SNP, we compute as many similarity scores
as haplotypes present in the reference. We have defined the similarity score as
the frequency of this genotype in the different haplotype population. To compute
the global similarity score, we have computed a mean of the scores by SNP weighted
by the R2 between the SNP and the haplotype classification.

In addition, we have computed the probability of a sample to belong to each of
the haplotypes. To do so, we compute the conditional probability of a sample genotype
to belong to each haplotype and we apply Bayes' theorem.
}

