% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifSNPsImputation.R
\name{classifSNPsImpute}
\alias{classifSNPsImpute}
\title{Get similarity scores and probability using imputed data.}
\usage{
classifSNPsImpute(genos, R2, refs, mc.cores = 1)
}
\arguments{
\item{genos}{Array with the genotypes probabilities.}

\item{R2}{Vector with the R2 between the SNPs and the inversion status}

\item{refs}{List of matrices. Each matrix has, for an SNP, the frequencies of each genotype in the
different haplotypes.}

\item{mc.cores}{Numeric with the number of cores used in the computation}
}
\value{
List with the results:
\itemize{
\item{scores: Matrix with the simmilarity scores of the individuals}
\item{numSNPs: Vector with the number of SNPs used in each computation}
}
}
\description{
This function computes the similarity scores between the sample SNPs and the haplotype's
reference. It also returns the probability of each sample to belong to the
different haplotypes.
}
\details{
This function computes, for each individual, similarity scores for
all the present haplotypes. For each SNP, we compute as many similarity scores
as haplotypes present in the reference. We have defined the similarity score as
the frequency of this genotype in the different haplotype population. To compute
the global similarity score, we have computed a mean of the scores by SNP weighted
by the R2 between the SNP and the haplotype classification.

In addition, we have computed the probability of a sample to belong to each of
the haplotypes. To do so, we compute the conditional probability of a sample genotype
to belong to each haplotype and we apply Bayes' theorem.
}
