% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifSNPs.R
\name{classifSNPspar}
\alias{classifSNPspar}
\title{Get similarity scores}
\usage{
classifSNPspar(genos, R2, refs, alfreq, genofreq, mc.cores)
}
\arguments{
\item{genos}{Matrix with the samples genotypes. It is the result of \code{getGenotypesTable}}

\item{R2}{Vector with the R2 between the SNPs and the inversion status}

\item{refs}{List of matrices. Each matrix has, for an SNP, the frequencies of each genotype in the
homozygous inverted, homozygous standard and heterozygous samples.}

\item{mc.cores}{Numeric with the number of cores used in the computation}
}
\value{
Matrix with the standard and inverted scores for each individual. Samples are in columns and
scores in rows
}
\description{
This function computes the similarity scores between the sample SNPs and the inversion's reference.
}
\details{
This function computes two similarity scores for each individual: the inverted and the
standard score. The individuals' score are computed as a weighted mean of the similarity scores
of each SNP.

For each SNP, we compare the genotype of the invidual with the genotypes frequencies
in the homozygous inverted and standard samples. If the frequency of the individual genotype is
higher in the inverted samples, we add 1 to the inverted score and 0 to the standard score. If the
frequency of the individual genotype is higher in the standard sample, we add 1 to the standard
score and 0 to the inverted score. The individuals' scores are computed using the mean of the
SNPs scores weigthed by the R2 between the SNP and the inversion status.
}

