% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifSNPs.R
\name{computeScore}
\alias{computeScore}
\title{Compute all similarity scores for a sample}
\usage{
computeScore(geno, refs, R2)
}
\arguments{
\item{geno}{Vector with the sample genotypes. It is the result of
\code{getGenotypesTable}}

\item{refs}{List of matrices. Each matrix has, for an SNP, the frequencies of each genotype in the
different haplotypes.}

\item{R2}{Vector with the R2 between the SNPs and the inversion status}
}
\value{
List with the results:
\itemize{
\item{scores: Vector with the simmilarity scores of the sample}
\item{numSNPs: Numeric with the number of SNPs used in the computation}
}
}
\description{
Internal
}
