% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctAlleleTable.R
\name{correctAlleleTable}
\alias{correctAlleleTable}
\title{Solve genotypes discrepancies}
\usage{
correctAlleleTable(alleletable, hetRefs, map)
}
\arguments{
\item{alleletable}{Data.frame with the alleles per SNP (from getAlleleTable)}

\item{hetRefs}{Character vector with the heterozygous genotypes in the reference.}

\item{map}{Data.frame with the annotation of the SNPs (from plink format)}
}
\value{
alleletable without discrepancies between these genotypes and the references.
}
\description{
This function tries to solve discrepancies between the reference and sample genotypes. The cause of
these discrepancies is that samples and references have used different strands to codify the SNP. This
function get the complement genotypes for the discordant SNPs and checks if discordancies are solved.
}

