% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlleleTable.R
\name{getAlleleTable}
\alias{getAlleleTable}
\title{Compute the allele table}
\usage{
getAlleleTable(map)
}
\arguments{
\item{map}{Data.frame with the annotation of the SNPs (from plink format)}
}
\value{
Data.frame with genotypes map
}
\description{
Get a data.frame that maps the numeric genotype of a SNPmatrix (0, 1, 2) into the real genotype.
Heterozygous genotypes are ordered alphabetically.
}

