% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenotypesTable.R
\name{getGenotypesTable}
\alias{getGenotypesTable}
\title{Get genotypes table}
\usage{
getGenotypesTable(geno, allele)
}
\arguments{
\item{geno}{SnpMatrix (from plink format)}

\item{allele}{Data.frame with the alleles per SNP (from getAlleleTable)}
}
\value{
Character matrix with the samples genotypes
}
\description{
Get a matrix with the sample genotypes from all SNP.
}

