% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalFreqs.R
\docType{data}
\name{globalFreqs}
\alias{globalFreqs}
\title{Frequency of each allele in European population}
\format{List of matrices for four inversions (ROIno.8.3, ROIno.17.16, HsInv0286 and
HsInv0396). Each matrix has the frequency of each genotype in the European population.}
\usage{
globalFreqs
}
\description{
Dataset with the genotype frequencies in the European population. These frequencies
have been computed using the European samples of 1000 Genomes Phase 3 data.
}
\keyword{datasets}

