% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetRefs.R
\docType{data}
\name{hetRefs}
\alias{hetRefs}
\title{Heterozygote genotypes in the references}
\format{Character vector with the heterozygous genotypes. The vector is named with the SNPs names.}
\usage{
hetRefs
}
\description{
Dataset with the heterozygote genotypes of all the SNPs used in any of the references. This dataset
include all the SNPs that are present inside the inversion's region in 1000 Genomes Phase 3.
}
\keyword{datasets}

