% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\docType{data}
\name{info}
\alias{info}
\title{SNP reference description}
\format{data.frame}
\usage{
info
}
\description{
Description of the SNPs inclued in scoreInvHap references. The description includes
the coordinates in hg19, the dbSNP identifier, the reference and alternative allele
and the allele frequency in the European Samples of 1000 Genomes.
}
\keyword{datasets}
