% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inversionGR.R
\docType{data}
\name{inversionGR}
\alias{inversionGR}
\title{Inversions' description}
\format{GenomicRanges with the inversions' description in the metada}
\usage{
inversionGR
}
\description{
Description of the 20 human inversions whose references are included in
scoreInvHap. The description includes the citogenic location, the coordinates in
hg19, the number of alleles and the number of SNPs with a MAD > 5% in the European
Samples of 1000 Genomes.
}
\keyword{datasets}
