% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreInvHap.R
\name{prepareMap}
\alias{prepareMap}
\title{Modify feature data from VCF}
\usage{
prepareMap(vcf)
}
\arguments{
\item{vcf}{\code{VCF} object}
}
\value{
Data.frame with the feature data
}
\description{
Internal. Modify feature data from VCF to comply with
scoreInvHap requirements.
}
