% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refs.R
\docType{data}
\name{Refs}
\alias{Refs}
\title{Genotype frequency in references}
\format{List of matrices for four inversions (inv8p23.1, inv17q21.31, inv7p11.2
and invXq13.2). Each matrices has the frequency of each genotype in each haplotype.}
\usage{
Refs
}
\description{
Dataset with the genotype frequencies in the different haplotype populations.
These frequencies have been computed using the European samples of 1000 Genomes
Phase 3 data. Real inversion status have been estimated using invClust.
}
\keyword{datasets}
