% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPfieRes_class.R
\docType{class}
\name{SNPfieRes}
\alias{SNPfieRes}
\alias{SNPfieRes-class}
\alias{SNPfieRes-methods}
\alias{classification,SNPfieRes-method}
\alias{SNPfieRes-methods}
\alias{classification}
\alias{certainty,SNPfieRes-method}
\alias{SNPfieRes-methods}
\alias{certainty}
\alias{numSNPs,SNPfieRes-method}
\alias{SNPfieRes-methods}
\alias{numSNPs}
\alias{scores,SNPfieRes-method}
\alias{SNPfieRes-methods}
\alias{scores}
\title{SNPfieRes instances}
\usage{
\S4method{classification}{SNPfieRes}(object)

\S4method{certainty}{SNPfieRes}(object)

\S4method{numSNPs}{SNPfieRes}(object)

\S4method{scores}{SNPfieRes}(object)
}
\arguments{
\item{object}{\code{SNPfieRes}}
}
\value{
A SNPfieRes instance
}
\description{
Container with the results of the classification pipeline
}
\section{Methods (by generic)}{
\itemize{
\item \code{classification}: Get classification

\item \code{certainty}: Get classification certainty

\item \code{numSNPs}: Get number of SNPs used in computation

\item \code{scores}: Get similarity scores
}}

\section{Slots}{

\describe{
\item{\code{classification}}{Factor with the individuals classification}

\item{\code{scores}}{Simmilarity scores for the different haplotypes.}

\item{\code{numSNPs}}{Numeric with SNPs used to compute the scores.}

\item{\code{certainty}}{Numeric with the certainty of the classification for each individual.}
}}

\examples{
if(require(VariantAnnotation)){
  vcf <- readVcf(system.file("extdata", "example.vcf", package = "snpfier"), "hg19")

  ## Create SNPfieRes class from pipeline
  res <- classifierPipeline(vcf, SNPsR2$HsInv0286, hetRefs = hetRefs$HsInv0286,
  Refs$HsInv0286, mc.cores = 1)

  ## Print object
  res

  ## Get haplotype classification
  classification(res)

  ## Get similiraty scores
  scores(res)
}
}
