% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPsR2.R
\docType{data}
\name{SNPsR2}
\alias{SNPsR2}
\title{R2 between the SNPs and the inversion status}
\format{List of numeric vectors for four inversions (inv8p23.1, inv17q21.31, inv7p11.2
and invXq13.2).}
\usage{
SNPsR2
}
\description{
Dataset with R2 between the SNPs and the inversion status. This values are used to weigth
similarity scores. These values have been computed using the European samples
of 1000 Genomes Phase 3 data. Real inversion status have been estimated using
invClust.
}
\keyword{datasets}
