% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierPipeline.R
\name{classifierPipeline}
\alias{classifierPipeline}
\title{Classify inversions using identity}
\usage{
classifierPipeline(SNPlist, SNPsR2, hetRefs, Refs, R2 = 0.3, alfreq, genofreq,
  mc.cores = 1, verbose = FALSE)
}
\arguments{
\item{SNPlist}{List with SNPs data. It should contain genotypes (a \code{SNPmatrix}) and map (a data.frame
with the annotation)}

\item{SNPsR2}{Vector with the R2 of the SNPs of the region}

\item{hetRefs}{Vector with the heterozygote form of the SNP in the inversion}

\item{Refs}{List with the allele frequencies in the references}

\item{R2}{Vector with the R2 between the SNPs and the inversion status}

\item{alfreq}{List with the allele frequencies in the target population}

\item{genofreq}{Vector with the frequency of each haplotype}

\item{mc.cores}{Numeric with the number of cores used in the computation}

\item{verbose}{Should message be shown?}
}
\value{
A \code{SNPfieRes} object
}
\description{
Classify inversions using identity
}

