% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierPipeline.R
\name{classifierPipeline}
\alias{classifierPipeline}
\title{Classify inversions using identity}
\usage{
classifierPipeline(SNPlist, SNPsR2, hetRefs, Refs, R2 = 0, imputed = FALSE,
  mc.cores = 1, verbose = FALSE)
}
\arguments{
\item{SNPlist}{List with SNPs data. It should contain genotypes (a \code{SNPmatrix}) and map (a data.frame
with the annotation)}

\item{SNPsR2}{Vector with the R2 of the SNPs of the region}

\item{hetRefs}{Vector with the heterozygote form of the SNP in the inversion}

\item{Refs}{List with the allele frequencies in the references}

\item{R2}{Vector with the R2 between the SNPs and the inversion status}

\item{imputed}{Logical. If TRUE, scores are computed using posterior probabilities.
If FALSE, scores are computed using best guess. Only applied when SNPlist is a VCF.}

\item{mc.cores}{Numeric with the number of cores used in the computation}

\item{verbose}{Should message be shown?}
}
\value{
A \code{SNPfieRes} object
}
\description{
This is the main function of `snpfier` package. This function accepts
SNPs data in a plink or a VCF format and compute the inversion prediction.
}
\examples{
if(require(VariantAnnotation)){
  vcf <- readVcf(system.file("extdata", "example.vcf", package = "snpfier"), "hg19")
  res <- classifierPipeline(vcf, SNPsR2$HsInv0286, hetRefs = hetRefs$HsInv0286,
  Refs$HsInv0286, mc.cores = 1)
}
}
