% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInvStatus.R
\name{getInvStatus}
\alias{getInvStatus}
\title{Get the inversion status of a sample}
\usage{
getInvStatus(scores)
}
\arguments{
\item{scores}{Matrix of similarity scores (from classifSNPspar)
to a group is smaller than this parameter, the individual is classified as NA.}
}
\value{
Factor with the inversion classification of the samples
}
\description{
This function estimates the inversion status of the samples using the similarity scores and a LDA
model.
}

