% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInvStatus.R
\name{getInvStatus}
\alias{getInvStatus}
\title{Get the inversion status of a sample}
\usage{
getInvStatus(scores)
}
\arguments{
\item{scores}{Matrix of probabilities (from \code{classifSNPs})}
}
\value{
List with the results:
\itemize{
\item{class: Vector with the most probable classification}
\item{certainty: Vector with the certainty of the most probable classification}
}
}
\description{
This function estimates the inversion status of the samples using the probabilities
computed in \code{classifSNPs}
}
