% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalFreqs.R
\docType{data}
\name{globalFreqs}
\alias{globalFreqs}
\title{Frequency of each allele in European population}
\format{List of character vectors with the heterozygous genotypes of the SNPs present included
the region of 6 inversions (ROIno.2.13, ROIno.3.2, ROIno.7.10, ROIno.8.3, ROIno.16.13
and ROIno.17.16). Each element is named with the SNPs names.}
\usage{
globalFreqs
}
\description{
Dataset with the heterozygote genotypes of all the SNPs used in any of the references. This dataset
include all the SNPs that are present inside the inversion's region in 1000 Genomes Phase 3.
}
\keyword{datasets}

