% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refs.R
\docType{data}
\name{Refs}
\alias{Refs}
\title{Genotype frequency in references}
\format{List of matrices for 6 inversions (ROIno.2.13, ROIno.3.2, ROIno.7.10, ROIno.8.3,
ROIno.16.13 and ROIno.17.16). Each matrices has the frequency of each genotype in each of the
inversion status}
\usage{
Refs
}
\description{
Dataset with the genotype frequencies in the different inversion samples (homozygous inverted,
heterozygous, homozygous inverted). These frequencies have been computed using the 1000 Genomes
Phase 3 data. Real inversion status have been estimated using invClust.
}
\keyword{datasets}

