% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPfieRes_class.R
\docType{class}
\name{SNPfieRes}
\alias{SNPfieRes}
\alias{SNPfieRes-class}
\alias{SNPfieRes-methods}
\alias{certainty}
\alias{certainty,SNPfieRes-method}
\alias{classification}
\alias{classification,SNPfieRes-method}
\alias{plot}
\alias{plot,SNPfieRes,ANY-method}
\alias{scores}
\alias{scores,SNPfieRes-method}
\title{SNPfieRes instances}
\usage{
\S4method{classification}{SNPfieRes}(object)

\S4method{certainty}{SNPfieRes}(object)

\S4method{scores}{SNPfieRes}(object)

\S4method{plot}{SNPfieRes,ANY}(x, y, ...)
}
\arguments{
\item{object}{\code{SNPfieRes}}

\item{object}{\code{SNPfieRes}}

\item{object}{\code{SNPfieRes}}

\item{object}{\code{SNPfieRes}}
}
\description{
Container with the results of the classification pipeline
}
\section{Methods (by generic)}{
\itemize{
\item \code{classification}: Get samples classification

\item \code{certainty}: Get classification certainty

\item \code{scores}: Get samples similarity scores

\item \code{plot}: Plot samples similarity scores
}}
\section{Slots}{

\describe{
\item{\code{classification}}{Factor with the individuals classification}

\item{\code{certainty}}{Numeric with the certainty of the classification for each individual.}

\item{\code{scores}}{Simmilarity scores for the inverted and standard references.}
}}

