% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPfieRes_class.R
\docType{class}
\name{SNPfieRes}
\alias{SNPfieRes}
\alias{SNPfieRes-class}
\alias{SNPfieRes-methods}
\alias{classification,SNPfieRes-method}
\alias{SNPfieRes-methods}
\alias{classification}
\alias{certainty,SNPfieRes-method}
\alias{SNPfieRes-methods}
\alias{certainty}
\alias{diffscores,SNPfieRes-method}
\alias{SNPfieRes-methods}
\alias{diffscores}
\alias{maxscores,SNPfieRes-method}
\alias{SNPfieRes-methods}
\alias{maxscores}
\alias{numSNPs,SNPfieRes-method}
\alias{SNPfieRes-methods}
\alias{numSNPs}
\alias{plotCallRate,SNPfieRes-method}
\alias{SNPfieRes-methods}
\alias{plotCallRate}
\alias{plotScores,SNPfieRes-method}
\alias{SNPfieRes-methods}
\alias{plotScores}
\alias{propSNPs,SNPfieRes-method}
\alias{SNPfieRes-methods}
\alias{propSNPs}
\alias{scores,SNPfieRes-method}
\alias{SNPfieRes-methods}
\alias{scores}
\title{SNPfieRes instances}
\usage{
\S4method{classification}{SNPfieRes}(object, minDiff = 0, callRate = 0,
  inversion = FALSE)

\S4method{certainty}{SNPfieRes}(object)

\S4method{diffscores}{SNPfieRes}(object)

\S4method{maxscores}{SNPfieRes}(object)

\S4method{numSNPs}{SNPfieRes}(object)

\S4method{plotCallRate}{SNPfieRes}(object, callRate = 0.9, ...)

\S4method{plotScores}{SNPfieRes}(object, minDiff = 0.1, ...)

\S4method{propSNPs}{SNPfieRes}(object)

\S4method{scores}{SNPfieRes}(object)
}
\arguments{
\item{object}{\code{SNPfieRes}}

\item{minDiff}{Numeric with the threshold of the minimum difference
between the top and the second score. Used to filter samples.}

\item{callRate}{Numeric with the threshold of the minimum call rate
of the samples. Used to filter samples.}

\item{inversion}{Logical. If true, haplotypes classification is adapted
to return inversion status.}

\item{...}{Further parameters passed to plot function.}
}
\value{
A SNPfieRes instance
}
\description{
Container with the results of the classification pipeline
}
\section{Methods (by generic)}{
\itemize{
\item \code{classification}: Get classification

\item \code{certainty}: Get classification certainty

\item \code{diffscores}: Get maximum similarity scores

\item \code{maxscores}: Get maximum similarity scores

\item \code{numSNPs}: Get number of SNPs used in computation

\item \code{plotCallRate}: Plot call rate based QC

\item \code{plotScores}: Plot scores based QC

\item \code{propSNPs}: Get proportions of SNPs used in computation

\item \code{scores}: Get similarity scores
}}

\section{Slots}{

\describe{
\item{\code{classification}}{Factor with the individuals classification}

\item{\code{scores}}{Simmilarity scores for the different haplotypes.}

\item{\code{numSNPs}}{Numeric with SNPs used to compute the scores.}

\item{\code{certainty}}{Numeric with the certainty of the classification for each individual.}
}}

\examples{
if(require(VariantAnnotation)){
  vcf <- readVcf(system.file("extdata", "example.vcf", package = "snpfier"), "hg19")

  ## Create SNPfieRes class from pipeline
  res <- classifierPipeline(vcf, SNPsR2$HsInv0286, hetRefs = hetRefs$HsInv0286,
  Refs$HsInv0286, mc.cores = 1)

  ## Print object
  res

  ## Get haplotype classification
  classification(res)

  ## Get similiraty scores
  scores(res)
}
}
