% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPfieRes_class.R
\docType{class}
\name{SNPfieRes}
\alias{SNPfieRes}
\alias{SNPfieRes-class}
\alias{SNPfieRes-methods}
\alias{certainty}
\alias{certainty,SNPfieRes-method}
\alias{classification}
\alias{classification,SNPfieRes-method}
\alias{numSNPs}
\alias{numSNPs,SNPfieRes-method}
\alias{probs}
\alias{probs,SNPfieRes-method}
\alias{scores}
\alias{scores,SNPfieRes-method}
\title{SNPfieRes instances}
\usage{
\S4method{classification}{SNPfieRes}(object)

\S4method{certainty}{SNPfieRes}(object)

\S4method{numSNPs}{SNPfieRes}(object)

\S4method{probs}{SNPfieRes}(object)

\S4method{scores}{SNPfieRes}(object)
}
\arguments{
\item{object}{\code{SNPfieRes}}
}
\description{
Container with the results of the classification pipeline
}
\section{Methods (by generic)}{
\itemize{
\item \code{classification}: Get samples classification

\item \code{certainty}: Get classification certainty

\item \code{numSNPs}: Get number of SNPs used in computation

\item \code{probs}: Get samples probabilities

\item \code{scores}: Get samples similarity scores
}}
\section{Slots}{

\describe{
\item{\code{classification}}{Factor with the individuals classification}

\item{\code{probs}}{Probability of each individual to belong to the different haplotypes}

\item{\code{scores}}{Simmilarity scores for the different haplotypes.}

\item{\code{numSNPs}}{Numeric with SNPs used to compute the scores.}

\item{\code{certainty}}{Numeric with the certainty of the classification for each individual.}
}}

